Bake Red Hat CA certificate into the image
In many CI jobs where this image is used, the first few steps before
running the test commands includes:

- download and install RH-IT-Root-CA cert
- set REQUESTS_CA_BUNDLE so that pip-installed requests will respect
  the system CA bundle

It doesn't make much sense to me to require all the users to do this.
There are a few issues with it:

- it's inconvenient.

- every time that cert is downloaded is a new vector for attack. If the
  download can be somehow hijacked, the subsequent commands can be
  tricked into trusting any service.

- some users are probably not going to investigate the situation
  properly and will instead disable SSL verification.

The cert is not a secret, so let's just bake it into the image.