Use Fedora31 and add Python2 packages for radas use
RADAS is still on Python2. It was using the legacy tagged image but
a new dependency needs to be added. Since the branch that created
legacy no longer exists, we need to create a new branch for the new
dependency. Since the newer Fedora versions stopped supporting python2,
we had to use Fedora31.